rename service id module so that it doesn't conflict with __init__
