+++ ansible_base/resource_registry/models/__init__.py
@@ -1,2 +1,2 @@
 from .resource import Resource, ResourceType, init_resource_from_object  # noqa: 401
-from .service_id import service_id  # noqa: 401
+from .service_identifier import service_id  # noqa: 401
