+++ ansible_base/resource_registry/models/service_identifier.py
